+++ tasks.rb
@@ -110,11 +110,13 @@ def create_hpxmls
   puts "\n"
 
   # Print warnings about extra files
-  dirs.each do |dir|
-    Dir["#{workflow_dir}/#{dir}/*.xml"].each do |hpxml|
-      next if abs_hpxml_files.include? File.absolute_path(hpxml)
+  if abs_hpxml_files.size == json_inputs.size
+    dirs.each do |dir|
+      Dir["#{workflow_dir}/#{dir}/*.xml"].each do |hpxml|
+        next if abs_hpxml_files.include? File.absolute_path(hpxml)
 
-      puts "Warning: Extra HPXML file found at #{File.absolute_path(hpxml)}"
+        puts "Warning: Extra HPXML file found at #{File.absolute_path(hpxml)}"
+      end
     end
   end
 end
@@ -295,7 +297,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     end
   end
   hpxml_bldg.roofs.each do |roof|
-    next unless roof.interior_adjacent_to == HPXML::LocationLivingSpace
+    next unless roof.interior_adjacent_to == HPXML::LocationConditionedSpace
 
     roof.interior_finish_type = HPXML::InteriorFinishGypsumBoard
   end
@@ -303,7 +305,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     if surface.is_a?(HPXML::FoundationWall) && surface.interior_adjacent_to != HPXML::LocationBasementConditioned
       surface.interior_finish_type = HPXML::InteriorFinishNone
     end
-    next unless [HPXML::LocationLivingSpace,
+    next unless [HPXML::LocationConditionedSpace,
                  HPXML::LocationBasementConditioned].include?(surface.interior_adjacent_to) &&
                 [HPXML::LocationOutside,
                  HPXML::LocationGround,
@@ -350,7 +352,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
       adjacent_to = HPXML::LocationOtherHousingUnit
     end
     wall = hpxml_bldg.walls.select { |w|
-             w.interior_adjacent_to == HPXML::LocationLivingSpace &&
+             w.interior_adjacent_to == HPXML::LocationConditionedSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit
            }[0]
     wall.exterior_adjacent_to = adjacent_to
@@ -379,13 +381,13 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.cooking_ranges[0].location = adjacent_to
   elsif ['base-bldgtype-multifamily-adjacent-to-multiple.xml'].include? hpxml_file
     wall = hpxml_bldg.walls.select { |w|
-             w.interior_adjacent_to == HPXML::LocationLivingSpace &&
+             w.interior_adjacent_to == HPXML::LocationConditionedSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit
            }[0]
     wall.delete
     hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                          exterior_adjacent_to: HPXML::LocationOtherHeatedSpace,
-                         interior_adjacent_to: HPXML::LocationLivingSpace,
+                         interior_adjacent_to: HPXML::LocationConditionedSpace,
                          wall_type: HPXML::WallTypeWoodStud,
                          area: 100,
                          solar_absorptance: 0.7,
@@ -394,7 +396,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          insulation_assembly_r_value: 23.0)
     hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                          exterior_adjacent_to: HPXML::LocationOtherMultifamilyBufferSpace,
-                         interior_adjacent_to: HPXML::LocationLivingSpace,
+                         interior_adjacent_to: HPXML::LocationConditionedSpace,
                          wall_type: HPXML::WallTypeWoodStud,
                          area: 100,
                          solar_absorptance: 0.7,
@@ -403,7 +405,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          insulation_assembly_r_value: 23.0)
     hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                          exterior_adjacent_to: HPXML::LocationOtherNonFreezingSpace,
-                         interior_adjacent_to: HPXML::LocationLivingSpace,
+                         interior_adjacent_to: HPXML::LocationConditionedSpace,
                          wall_type: HPXML::WallTypeWoodStud,
                          area: 100,
                          solar_absorptance: 0.7,
@@ -412,7 +414,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          insulation_assembly_r_value: 23.0)
     hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                          exterior_adjacent_to: HPXML::LocationOtherHousingUnit,
-                         interior_adjacent_to: HPXML::LocationLivingSpace,
+                         interior_adjacent_to: HPXML::LocationConditionedSpace,
                          wall_type: HPXML::WallTypeWoodStud,
                          area: 100,
                          solar_absorptance: 0.7,
@@ -424,27 +426,27 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.floors[0].insulation_id = 'Floor1Insulation'
     hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                           exterior_adjacent_to: HPXML::LocationOtherNonFreezingSpace,
-                          interior_adjacent_to: HPXML::LocationLivingSpace,
+                          interior_adjacent_to: HPXML::LocationConditionedSpace,
                           floor_type: HPXML::FloorTypeWoodFrame,
                           area: 550,
                           insulation_assembly_r_value: 18.7,
                           floor_or_ceiling: HPXML::FloorOrCeilingFloor)
     hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                           exterior_adjacent_to: HPXML::LocationOtherMultifamilyBufferSpace,
-                          interior_adjacent_to: HPXML::LocationLivingSpace,
+                          interior_adjacent_to: HPXML::LocationConditionedSpace,
                           floor_type: HPXML::FloorTypeWoodFrame,
                           area: 200,
                           insulation_assembly_r_value: 18.7,
                           floor_or_ceiling: HPXML::FloorOrCeilingFloor)
     hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                           exterior_adjacent_to: HPXML::LocationOtherHeatedSpace,
-                          interior_adjacent_to: HPXML::LocationLivingSpace,
+                          interior_adjacent_to: HPXML::LocationConditionedSpace,
                           floor_type: HPXML::FloorTypeWoodFrame,
                           area: 150,
                           insulation_assembly_r_value: 2.1,
                           floor_or_ceiling: HPXML::FloorOrCeilingFloor)
     wall = hpxml_bldg.walls.select { |w|
-             w.interior_adjacent_to == HPXML::LocationLivingSpace &&
+             w.interior_adjacent_to == HPXML::LocationConditionedSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherMultifamilyBufferSpace
            }[0]
     hpxml_bldg.windows.add(id: "Window#{hpxml_bldg.windows.size + 1}",
@@ -455,7 +457,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                            fraction_operable: 0.67,
                            wall_idref: wall.id)
     wall = hpxml_bldg.walls.select { |w|
-             w.interior_adjacent_to == HPXML::LocationLivingSpace &&
+             w.interior_adjacent_to == HPXML::LocationConditionedSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherHeatedSpace
            }[0]
     hpxml_bldg.doors.add(id: "Door#{hpxml_bldg.doors.size + 1}",
@@ -464,7 +466,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          azimuth: 0,
                          r_value: 4.4)
     wall = hpxml_bldg.walls.select { |w|
-             w.interior_adjacent_to == HPXML::LocationLivingSpace &&
+             w.interior_adjacent_to == HPXML::LocationConditionedSpace &&
                w.exterior_adjacent_to == HPXML::LocationOtherHousingUnit
            }[0]
     hpxml_bldg.doors.add(id: "Door#{hpxml_bldg.doors.size + 1}",
@@ -515,7 +517,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     end
     hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                          exterior_adjacent_to: HPXML::LocationAtticUnvented,
-                         interior_adjacent_to: HPXML::LocationLivingSpace,
+                         interior_adjacent_to: HPXML::LocationConditionedSpace,
                          wall_type: HPXML::WallTypeWoodStud,
                          area: 316,
                          solar_absorptance: 0.7,
@@ -524,7 +526,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          insulation_assembly_r_value: 23.0)
     hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                          exterior_adjacent_to: HPXML::LocationOutside,
-                         interior_adjacent_to: HPXML::LocationLivingSpace,
+                         interior_adjacent_to: HPXML::LocationConditionedSpace,
                          wall_type: HPXML::WallTypeWoodStud,
                          siding: HPXML::SidingTypeWood,
                          area: 240,
@@ -547,7 +549,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     end
     hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                           exterior_adjacent_to: HPXML::LocationAtticUnvented,
-                          interior_adjacent_to: HPXML::LocationLivingSpace,
+                          interior_adjacent_to: HPXML::LocationConditionedSpace,
                           floor_type: HPXML::FloorTypeWoodFrame,
                           area: 450,
                           interior_finish_type: HPXML::InteriorFinishGypsumBoard,
@@ -692,7 +694,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.rim_joists << hpxml_bldg.rim_joists[-1].dup
     hpxml_bldg.rim_joists[-1].id = "RimJoist#{hpxml_bldg.rim_joists.size}"
     hpxml_bldg.rim_joists[-1].insulation_id = "RimJoist#{hpxml_bldg.rim_joists.size}Insulation"
-    hpxml_bldg.rim_joists[-1].interior_adjacent_to = HPXML::LocationLivingSpace
+    hpxml_bldg.rim_joists[-1].interior_adjacent_to = HPXML::LocationConditionedSpace
     hpxml_bldg.rim_joists[-1].area = 116
   elsif ['base-foundation-conditioned-basement-wall-insulation.xml'].include? hpxml_file
     hpxml_bldg.foundation_walls.each do |foundation_wall|
@@ -795,7 +797,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.floors[0].area = 675
     hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                           exterior_adjacent_to: HPXML::LocationCrawlspaceUnvented,
-                          interior_adjacent_to: HPXML::LocationLivingSpace,
+                          interior_adjacent_to: HPXML::LocationConditionedSpace,
                           floor_type: HPXML::FloorTypeWoodFrame,
                           area: 675,
                           insulation_assembly_r_value: 18.7,
@@ -915,7 +917,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                          insulation_assembly_r_value: 4)
     hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                           exterior_adjacent_to: HPXML::LocationGarage,
-                          interior_adjacent_to: HPXML::LocationLivingSpace,
+                          interior_adjacent_to: HPXML::LocationConditionedSpace,
                           floor_type: HPXML::FloorTypeWoodFrame,
                           area: 400,
                           insulation_assembly_r_value: 39.3,
@@ -955,7 +957,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     floors_map.each_with_index do |(floor_type, assembly_r), _i|
       hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                             exterior_adjacent_to: exterior_adjacent_to,
-                            interior_adjacent_to: HPXML::LocationLivingSpace,
+                            interior_adjacent_to: HPXML::LocationConditionedSpace,
                             floor_type: floor_type,
                             area: area / floors_map.size,
                             insulation_assembly_r_value: assembly_r,
@@ -974,7 +976,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     floors_map.each_with_index do |(floor_type, assembly_r), _i|
       hpxml_bldg.floors.add(id: "Floor#{hpxml_bldg.floors.size + 1}",
                             exterior_adjacent_to: exterior_adjacent_to,
-                            interior_adjacent_to: HPXML::LocationLivingSpace,
+                            interior_adjacent_to: HPXML::LocationConditionedSpace,
                             floor_type: floor_type,
                             area: area / floors_map.size,
                             insulation_assembly_r_value: assembly_r,
@@ -1042,7 +1044,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     walls_map.each_with_index do |(wall_type, assembly_r), i|
       hpxml_bldg.walls.add(id: "Wall#{hpxml_bldg.walls.size + 1}",
                            exterior_adjacent_to: HPXML::LocationOutside,
-                           interior_adjacent_to: HPXML::LocationLivingSpace,
+                           interior_adjacent_to: HPXML::LocationConditionedSpace,
                            wall_type: wall_type,
                            siding: siding_types[i % siding_types.size][0],
                            color: siding_types[i % siding_types.size][1],
@@ -1384,7 +1386,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.hvac_distributions[0].ducts[-1].id = "Ducts#{hpxml_bldg.hvac_distributions[0].ducts.size}"
     hpxml_bldg.hvac_distributions[0].ducts[2].duct_location = HPXML::LocationExteriorWall
     hpxml_bldg.hvac_distributions[0].ducts[2].duct_surface_area = 37.5
-    hpxml_bldg.hvac_distributions[0].ducts[3].duct_location = HPXML::LocationLivingSpace
+    hpxml_bldg.hvac_distributions[0].ducts[3].duct_location = HPXML::LocationConditionedSpace
     hpxml_bldg.hvac_distributions[0].ducts[3].duct_surface_area = 12.5
   elsif ['base-hvac-ducts-effective-rvalue.xml'].include? hpxml_file
     hpxml_bldg.hvac_distributions[0].ducts[0].duct_insulation_r_value = nil
@@ -1624,7 +1626,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                                  energy_factor: 1.6,
                                  rh_setpoint: 0.5,
                                  fraction_served: 0.25,
-                                 location: HPXML::LocationLivingSpace)
+                                 location: HPXML::LocationConditionedSpace)
   end
   if ['base-hvac-air-to-air-heat-pump-var-speed-backup-furnace.xml',
       'base-hvac-autosize-air-to-air-heat-pump-var-speed-backup-furnace.xml'].include? hpxml_file
@@ -1693,7 +1695,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                                          number_of_units_served: 6,
                                          fuel_type: HPXML::FuelTypeNaturalGas,
                                          water_heater_type: HPXML::WaterHeaterTypeStorage,
-                                         location: HPXML::LocationLivingSpace,
+                                         location: HPXML::LocationConditionedSpace,
                                          tank_volume: 120,
                                          fraction_dhw_load_served: 1.0,
                                          heating_capacity: 40000,
@@ -1717,7 +1719,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.water_heating_systems.add(id: "WaterHeatingSystem#{hpxml_bldg.water_heating_systems.size + 1}",
                                          fuel_type: HPXML::FuelTypeNaturalGas,
                                          water_heater_type: HPXML::WaterHeaterTypeStorage,
-                                         location: HPXML::LocationLivingSpace,
+                                         location: HPXML::LocationConditionedSpace,
                                          tank_volume: 50,
                                          fraction_dhw_load_served: 0.2,
                                          heating_capacity: 40000,
@@ -1727,7 +1729,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.water_heating_systems.add(id: "WaterHeatingSystem#{hpxml_bldg.water_heating_systems.size + 1}",
                                          fuel_type: HPXML::FuelTypeElectricity,
                                          water_heater_type: HPXML::WaterHeaterTypeHeatPump,
-                                         location: HPXML::LocationLivingSpace,
+                                         location: HPXML::LocationConditionedSpace,
                                          tank_volume: 80,
                                          fraction_dhw_load_served: 0.2,
                                          energy_factor: 2.3,
@@ -1735,20 +1737,20 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.water_heating_systems.add(id: "WaterHeatingSystem#{hpxml_bldg.water_heating_systems.size + 1}",
                                          fuel_type: HPXML::FuelTypeElectricity,
                                          water_heater_type: HPXML::WaterHeaterTypeTankless,
-                                         location: HPXML::LocationLivingSpace,
+                                         location: HPXML::LocationConditionedSpace,
                                          fraction_dhw_load_served: 0.2,
                                          energy_factor: 0.99,
                                          temperature: 125.0)
     hpxml_bldg.water_heating_systems.add(id: "WaterHeatingSystem#{hpxml_bldg.water_heating_systems.size + 1}",
                                          fuel_type: HPXML::FuelTypeNaturalGas,
                                          water_heater_type: HPXML::WaterHeaterTypeTankless,
-                                         location: HPXML::LocationLivingSpace,
+                                         location: HPXML::LocationConditionedSpace,
                                          fraction_dhw_load_served: 0.1,
                                          energy_factor: 0.82,
                                          temperature: 125.0)
     hpxml_bldg.water_heating_systems.add(id: "WaterHeatingSystem#{hpxml_bldg.water_heating_systems.size + 1}",
                                          water_heater_type: HPXML::WaterHeaterTypeCombiStorage,
-                                         location: HPXML::LocationLivingSpace,
+                                         location: HPXML::LocationConditionedSpace,
                                          tank_volume: 50,
                                          fraction_dhw_load_served: 0.1,
                                          related_hvac_idref: 'HeatingSystem1',
@@ -2027,7 +2029,7 @@ def apply_hpxml_modification(hpxml_file, hpxml)
                                    primary_indicator: false)
     end
     hpxml_bldg.freezers.add(id: "Freezer#{hpxml_bldg.freezers.size + 1}",
-                            location: HPXML::LocationLivingSpace,
+                            location: HPXML::LocationConditionedSpace,
                             rated_annual_kwh: 400)
     if hpxml_file == 'base-misc-usage-multiplier.xml'
       hpxml_bldg.freezers[-1].usage_multiplier = 0.9
@@ -2045,12 +2047,12 @@ def apply_hpxml_modification(hpxml_file, hpxml)
     hpxml_bldg.pools[0].heater_weekday_fractions = '0.003, 0.003, 0.003, 0.004, 0.008, 0.015, 0.026, 0.044, 0.084, 0.121, 0.127, 0.121, 0.120, 0.090, 0.075, 0.061, 0.037, 0.023, 0.013, 0.008, 0.004, 0.003, 0.003, 0.003'
     hpxml_bldg.pools[0].heater_weekend_fractions = '0.003, 0.003, 0.003, 0.004, 0.008, 0.015, 0.026, 0.044, 0.084, 0.121, 0.127, 0.121, 0.120, 0.090, 0.075, 0.061, 0.037, 0.023, 0.013, 0.008, 0.004, 0.003, 0.003, 0.003'
     hpxml_bldg.pools[0].heater_monthly_multipliers = '1.154, 1.161, 1.013, 1.010, 1.013, 0.888, 0.883, 0.883, 0.888, 0.978, 0.974, 1.154'
-    hpxml_bldg.hot_tubs[0].pump_weekday_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
-    hpxml_bldg.hot_tubs[0].pump_weekend_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
-    hpxml_bldg.hot_tubs[0].pump_monthly_multipliers = '0.837, 0.835, 1.084, 1.084, 1.084, 1.096, 1.096, 1.096, 1.096, 0.931, 0.925, 0.837'
-    hpxml_bldg.hot_tubs[0].heater_weekday_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
-    hpxml_bldg.hot_tubs[0].heater_weekend_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
-    hpxml_bldg.hot_tubs[0].heater_monthly_multipliers = '0.921, 0.928, 0.921, 0.915, 0.921, 1.160, 1.158, 1.158, 1.160, 0.921, 0.915, 0.921'
+    hpxml_bldg.permanent_spas[0].pump_weekday_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
+    hpxml_bldg.permanent_spas[0].pump_weekend_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
+    hpxml_bldg.permanent_spas[0].pump_monthly_multipliers = '0.837, 0.835, 1.084, 1.084, 1.084, 1.096, 1.096, 1.096, 1.096, 0.931, 0.925, 0.837'
+    hpxml_bldg.permanent_spas[0].heater_weekday_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
+    hpxml_bldg.permanent_spas[0].heater_weekend_fractions = '0.024, 0.029, 0.024, 0.029, 0.047, 0.067, 0.057, 0.024, 0.024, 0.019, 0.015, 0.014, 0.014, 0.014, 0.024, 0.058, 0.126, 0.122, 0.068, 0.061, 0.051, 0.043, 0.024, 0.024'
+    hpxml_bldg.permanent_spas[0].heater_monthly_multipliers = '0.921, 0.928, 0.921, 0.915, 0.921, 1.160, 1.158, 1.158, 1.160, 0.921, 0.915, 0.921'
   end
   if ['base-bldgtype-multifamily-shared-laundry-room.xml',
       'base-bldgtype-multifamily-shared-laundry-room-multiple-water-heaters.xml'].include? hpxml_file